Implemented issue #9

+++ src/environment/terrain/terrain.cpp
@@ -14,7 +14,7 @@ Terrain::Terrain(std::string type, const float width, const float height) :
 		growthProfiles[i] = std::make_unique<GrowthProfileQuadratic>(
 			5,
 			10,
-			(GROWTH_PROFILE_RESOLUTION - i) * 0.5f);
+			static_cast<size_t>((GROWTH_PROFILE_RESOLUTION - i) * 0.5f));
 }
 
 const std::string& Terrain::getType() const {
